Do not reset user counters if data exists

+++ public/reset.php
@@ -28,31 +28,44 @@ if (!empty($user)) {
     $userRef = MongoDBRef::create(User::getCollectionName(), $user->getId());
     $query = array('user' => $userRef);
     
-    $comments = Comment::all($query)->limit(100);
-    foreach ($comments as $comment) {
-        $start = microtime(true);
-        $comment->delete();
-        echo 'comment:' . $comment->getId() . " (" . runtime($start) .  "ms)\n";               
+    $total = Comment::all($query)->count();
+    if ($total > 0) {
+        $comments = Comment::all($query)->limit(100);
+        foreach ($comments as $comment) {
+            $start = microtime(true);
+            $comment->delete();
+            echo 'comment:' . $comment->getId() . " (" . runtime($start) .  "ms)\n";               
+        }
+    } else {
+        $user->totalComments = 0;
+        $user->save();
+    }    
+    
+    $total = Post::all($query)->count();
+    if ($total > 0) {
+        $posts = Post::all($query)->limit(100);
+        foreach ($posts as $post) {
+            $start = microtime(true);
+            $post->delete();
+            echo 'post:' . $post->getId() . " (" . runtime($start) .  "ms)\n";    
+        }
+    } else {
+        $user->totalPosts = 0;
+        $user->save();
     }
     
-    $posts = Post::all($query)->limit(100);
-    foreach ($posts as $post) {
-        $start = microtime(true);
-        $post->delete();
-        echo 'post:' . $post->getId() . " (" . runtime($start) .  "ms)\n";    
+    $total = Topic::all($query)->count();
+    if ($total > 0) {
+        $topics = Topic::all($query)->limit(100);
+        foreach ($topics as $topic) {
+            $start = microtime(true);
+            $topic->delete();
+            echo 'topic:' . $topic->getId() . " (" . runtime($start) .  "ms)\n";    
+        }
+    } else {
+        $user->totalTopics = 0;
+        $user->save();
     }
-    
-    $topics = Topic::all($query)->limit(100);
-    foreach ($topics as $topic) {
-        $start = microtime(true);
-        $topic->delete();
-        echo 'topic:' . $topic->getId() . " (" . runtime($start) .  "ms)\n";    
-    }
-    
-    $user->totalComments = 0;
-    $user->totalPosts = 0;
-    $user->totalTopics = 0;
-    $user->save();
     
     print "runtime: " . runtime() . " ms\n";
